Add about page content
Added a question mark to the header. Clicking on the logo in the header still links to CSU, but clicking on the team name will switch between the about page and the rest of the application.

The only content on the about page is a placeholder 'T## Team Name' and a close button to switch back to the app. Clicking on the header will still switch back as well.