Add constants.js and small fixes
Removed question mark from the header and commented out the vertical line in the css for the header.

Created Constants.js file and filled it with initial contents. Replaced hardcoded uses of team name in the client with the constant from Constants.js. I'm not sure where to place Constants.js, but we can discuss this with our directory structure refactor.

Replaced the logo link from CSU to ~314.