Changed zoom/addMarker behavior
Turns out that you can't mix and match click and mouse{up|down} events without avoiding strange behavior.
I originally tried to have a click event on the marker, and mouseup and mousedown events on the map. But when doing that, both events end up firing, when you would only want the marker event to fire and not the map event following a click on the marker.

Now I'm only using click events on both the map and marker.  So when you click on a marker, the callback function linked to the marker fires, and that's it (the click callback on the map does NOT subsequently fire as desired).  This means we can click on a marker to toggle the popup without also simulateously (and undersirably) adding a new marker (which is handled by the map click callback).

Now a single click on the map (where a marker is not present) does the following:
- add a marker to clicked location without adjusting the map center or zoom

Clicking on a marker does the following:
- displays the popup (to show the lat/lng coordinates)
- zooms in to level 10 (chosen because it's not too close and not too far)
- centers the map on the marker location
- BUT if the current zoom level is greater than 10 (i.e., we're zoomed in more than a 10 level), then we do not reset the zoomLevel to 10, it just stays as is.

I think this is better than zooming in AND centering on EVERY single click.
And I removed the "long click" (mouseup and mousedown events occurs >250 milliseconds apart) to avoid the weird behavior mentioned above.