Change Ulog to use URL Parameters
Ulog has most of the behavior for using URL parameters in place by
default. However, their default log level is warn, but we want error to
be our default. There is no way to change this default, and if we try to
overwrite the log level, it will ignore the url paramater entirely. So I
only set the log level if it isn't set.

I also removed all the log level behavior in place previously as well.